refactor activities and use more fragments, add drawer support
