+++ app/src/main/java/org/otfusion/caturday/application/ApplicationModule.java
@@ -71,7 +71,7 @@ public class ApplicationModule {
     @Provides
     @Singleton
     public Bus provideBus() {
-        return new Bus();
+        return new Bus("varonil");
     }
 
 }
