+++ app/src/main/java/org/otfusion/caturday/ui/activities/CatActivity.java
@@ -1,5 +1,6 @@
 package org.otfusion.caturday.ui.activities;
 
+import android.app.FragmentTransaction;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.ContextMenu;
@@ -7,8 +8,10 @@ import android.widget.AdapterView;
 
 import com.squareup.otto.Bus;
 
+import org.otfusion.caturday.R;
 import org.otfusion.caturday.application.VoteCatsApplication;
 import org.otfusion.caturday.service.CatService;
+import org.otfusion.caturday.ui.fragments.BaseFragment;
 import org.otfusion.caturday.util.ApplicationUtils;
 
 import javax.inject.Inject;
@@ -41,6 +44,16 @@ public abstract class CatActivity extends AppCompatActivity {
         setContentView(getContentLayoutId());
         ButterKnife.bind(this);
         loadUIContent();
+    }
+
+    protected void startFragment(BaseFragment fragment) {
+        startFragment(fragment, null);
+    }
+
+    protected void startFragment(BaseFragment fragment, String tag) {
+        FragmentTransaction fragmentTransaction = getFragmentManager().beginTransaction();
+        fragmentTransaction.add(R.id.fragment_container, fragment, tag);
+        fragmentTransaction.commit();
     }
 
     protected Bus getBus() {
