+++ app/src/main/java/org/otfusion/caturday/ui/framework/drawer/Drawer.java
@@ -1,0 +1,8 @@
+package org.otfusion.caturday.ui.framework.drawer;
+
+import android.view.View;
+
+public interface Drawer {
+
+    void init(View view);
+}
