+++ app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,43 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:layout_marginTop="?attr/actionBarSize">
+
+    <ImageView
+        android:id="@+id/cat_view"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"
+        android:layout_marginBottom="16dp"
+        android:layout_marginTop="?attr/actionBarSize"
+        android:layout_gravity="center"/>
+
+    <Button
+        android:id="@+id/favorite_cat_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="right|top"
+        style="@style/AppTheme"
+        android:elevation="100dp"
+        android:text="Favorite Cat"/>
+
+    <Button
+        android:id="@+id/load_cat_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        style="@style/AppTheme"
+        android:elevation="100dp"
+        android:text="Load Cat"/>
+
+    <Button
+        android:id="@+id/share_cat_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal|top"
+        style="@style/AppTheme"
+        android:elevation="100dp"
+        android:text="Share" />
+
+</FrameLayout>
